[Refactor] Using DBConnection object
Abstracted DB connection logic into a class. This simplifies the other db objects.
