+++ webserver/src/player_db.py
@@ -3,26 +3,12 @@
 import os
 import psycopg2
 
+from db_connection import DBConnection
+
 class PlayerDBConnection:
     def __init__(self):
-
-        self.con = None
-        self.config  = {
-            'db_host': os.environ['DB_PORT_5432_TCP_ADDR'],
-            'db_port': os.environ['DB_PORT_5432_TCP_PORT'],
-            'db_pass': os.environ['DB_PASSWORD']
-        }
-        try:
-             self.con = psycopg2.connect(
-                            database='docker',
-                            user='docker',
-                            host=self.config['db_host'],
-                            port=self.config['db_port'],
-                            password=self.config['db_pass'])
-        except psycopg2.Error as e:
-            print 'psycopg Error %s' % e     
-        except Exception as e:
-            print 'Error %s' % e
+        self.db_conn            = DBConnection()
+        self.con                = self.db_conn.con
 
     def usernameExists(self, username):
         try:
@@ -47,6 +33,3 @@ class PlayerDBConnection:
             print 'Database Error %s' % e
             return e
 
-    def __del__(self):
-        if self.con:
-            self.con.close()
