+++ webserver/src/tournament_db.py
@@ -3,26 +3,13 @@
 import os
 import psycopg2
 
+from db_connection import DBConnection
+
 class TournamentDBConnection:
     def __init__(self):
+        self.db_conn            = DBConnection()
+        self.con                = self.db_conn.con
 
-        self.con = None
-        self.config  = {
-            'db_host': os.environ['DB_PORT_5432_TCP_ADDR'],
-            'db_port': os.environ['DB_PORT_5432_TCP_PORT'],
-            'db_pass': os.environ['DB_PASSWORD']
-        }
-        try:
-             self.con = psycopg2.connect(
-                            database='docker',
-                            user='docker',
-                            host=self.config['db_host'],
-                            port=self.config['db_port'],
-                            password=self.config['db_pass'])
-        except psycopg2.Error as e:
-            print 'psycopg Error %s' % e     
-        except Exception as e:
-            print 'Error %s' % e
 
     def tournamentExists(self, name):
         try:
@@ -54,7 +41,3 @@ class TournamentDBConnection:
         except psycopg2.DatabaseError as e:
             print 'Database Error %s' % e
             return e
-
-    def __del__(self):
-        if self.con:
-            self.con.close()
